fix: tranform sidemenu and content into component
