+++ src/components/MyAccount/styles.ts
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+	padding: 0rem 1rem;
+
+	.ant-layout {
+		background-color: rgba(0, 0, 0, 0);
+	}
+	.ant-layout-sider {
+		background-color: rgba(0, 0, 0, 0);
+	}
+	.ant-menu {
+		background-color: rgba(0, 0, 0, 0);
+	}
+`;
